refact: extract telemetry from discovery into new package
