+++ portal/discovery/trace.go
@@ -5,7 +5,7 @@ package discovery
 // SelectMultiHopWithTrace on MOLSRelayPolicy and by the matching siblings on
 // RelaySet, and consumed by:
 //
-//   - portal/discovery/metrics.go — emits low-cardinality fields to Prometheus
+//   - portal/telemetry/metrics.go — emits low-cardinality fields to Prometheus
 //     (no per-client labels: ClientHash never becomes a metric label).
 //   - sampled debug logs (zerolog) — ClientHash carried; LocalAddress
 //     intentionally NOT carried in the trace (PII-leak surface; ClientHash is
@@ -17,65 +17,3 @@ package discovery
 // See /home/alpha/.claude/plans/sophisticate-and-rationalize-discovery-rosy-parnas.md
 // (Phase 1 — Telemetry only) for the rationale.
 
-import "time"
-
-// SelectionTrace captures the inputs and outputs of one selection invocation.
-// All fields are populated by the *WithTrace methods; downstream consumers
-// (metrics emitter, debug logger) read but never mutate the trace.
-type SelectionTrace struct {
-	Timestamp time.Time
-
-	// ClientHash is hashToGF64(LocalAddress) — a single byte derived from the
-	// client identity. Used only for sampled debug-log correlation. Not a
-	// Prometheus label (would unbounded cardinality) and not a public field on
-	// any external API.
-	ClientHash uint8
-
-	// Mode is "priority" or "multihop", matching the calling method.
-	Mode string
-
-	// Pool snapshot at selection time.
-	PoolTotal    int
-	PoolEligible int
-	PoolFallback int
-
-	// Suppressed lists URLs excluded from selection along with the reason map.
-	Suppressed []string
-	Reasons    map[string]string
-
-	// Congested is true when the existing congestion-grid switch is active
-	// (RTT mean > molsCongestionRTTThreshold).
-	Congested bool
-
-	// NonLinear is true when the variant-grid multiplier flip is active
-	// (CV > molsCVThreshold).
-	NonLinear bool
-
-	// M1, M2 are the MOLS multipliers used for this selection.
-	M1, M2 uint8
-
-	// AvgRTT is the mean discovery RTT across the auto pool sample.
-	AvgRTT time.Duration
-
-	// CV is the coefficient of variation of per-relay discovery RTTs.
-	CV float64
-
-	// Ranked carries per-relay scoring detail for every candidate considered
-	// (excluded URLs appear in Suppressed/Reasons instead).
-	Ranked []TraceEntry
-
-	// OutputURLs is the final ordered list returned by the selection method.
-	OutputURLs []string
-
-	// SelectionTook is wall time spent in the selection method.
-	SelectionTook time.Duration
-}
-
-// TraceEntry captures per-relay scoring detail within a SelectionTrace.
-type TraceEntry struct {
-	URL       string
-	Score     int
-	Confirmed bool
-	RTT       time.Duration
-	Demoted   bool
-}
